requirements.txt
agno
openai
python-dotenv
pypdf
lancedb
sentence-transformers
requests
python-docx